move services and normalizers to separate entry point
