+++ libs/domain/cart/services/src/adapter/normalizers/model.ts
@@ -1,5 +1,5 @@
+import { ApiCartModel, CartId } from '@spryker-oryx/cart';
 import { CamelCase } from '@spryker-oryx/core/utilities';
-import { ApiCartModel } from '../../../models';
 
 export type DeserializedCartIncludes = {
   [P in ApiCartModel.Includes as `${CamelCase<P>}`]?: P extends
@@ -8,3 +8,12 @@ export type DeserializedCartIncludes = {
     ? ApiCartModel.Entry[]
     : never;
 };
+
+export type DeserializedCart = ApiCartModel.Attributes &
+  Pick<
+    DeserializedCartIncludes,
+    CamelCase<
+      ApiCartModel.Includes.Items | ApiCartModel.Includes.GuestCartItems
+    >
+  > &
+  CartId;
