+++ libs/domain/cart/services/src/totals/index.ts
@@ -1,0 +1,1 @@
+export * from './default-totals.service';
