+++ libs/domain/cart/src/index.ts
@@ -4,3 +4,4 @@ export * from './feature';
 export * from './mixins';
 export * from './models';
 export * from './services';
+export * from './services-reexports';
