+++ libs/domain/cart/src/services/adapter/index.ts
@@ -1,3 +1,1 @@
 export * from './cart.adapter';
-export * from './default-cart.adapter';
-export * from './normalizers';
