--- libs/domain/cart/src/services/adapter/normalizers/cart/carts.normalizer.ts
@@ -1,32 +1,0 @@
-import { Transformer, TransformerService } from '@spryker-oryx/core';
-import { Provider } from '@spryker-oryx/di';
-import { Observable, combineLatest, of } from 'rxjs';
-import { Cart } from '../../../../models';
-import { CartNormalizer } from './cart.normalizer';
-import { DeserializedCart } from './model';
-
-export const CartsNormalizer = 'oryx.CartsNormalizer*';
-
-export function cartsItemsNormalizer(
-  data: DeserializedCart[],
-  transformer: TransformerService
-): Observable<Cart[]> {
-  return data.length
-    ? combineLatest(
-        data.map((cart) => transformer.transform(cart, CartNormalizer))
-      )
-    : of([]);
-}
-
-export const cartsNormalizer: Provider[] = [
-  {
-    provide: CartsNormalizer,
-    useValue: cartsItemsNormalizer,
-  },
-];
-
-declare global {
-  interface InjectionTokensContractMap {
-    [CartsNormalizer]: Transformer<Cart[]>[];
-  }
-}
