--- libs/domain/cart/src/services/adapter/normalizers/cart/model.ts
@@ -1,12 +1,0 @@
-import { CamelCase } from '@spryker-oryx/core/utilities';
-import { ApiCartModel, CartId } from '../../../../models';
-import { DeserializedCartIncludes } from '../model';
-
-export type DeserializedCart = ApiCartModel.Attributes &
-  Pick<
-    DeserializedCartIncludes,
-    CamelCase<
-      ApiCartModel.Includes.Items | ApiCartModel.Includes.GuestCartItems
-    >
-  > &
-  CartId;
