--- libs/domain/cart/src/services/adapter/normalizers/index.ts
@@ -1,1 +1,0 @@
-export * from './cart';
