+++ libs/domain/cart/src/services/index.ts
@@ -1,7 +1,6 @@
 export * from './adapter';
 export * from './cart.providers';
 export * from './cart.service';
-export * from './default-cart.service';
 export * from './resolver';
 export * from './state';
 export * from './totals';
