+++ libs/domain/cart/src/services/totals/index.ts
@@ -1,3 +1,2 @@
 export * from './cart-totals.resolver';
-export * from './default-totals.service';
 export * from './totals.service';
